interface for common visualization made. TEsting left.
